+++ Source/SmithUE/Private/SmithUEModule.cpp
@@ -37,6 +37,7 @@
 #include "Dialog/SmithUEDialogWatcher.h"
 #include "Commands/SmithUEPCGCommands.h"
 #include "Commands/SmithUEAICommands.h"
+#include "Commands/SmithUEInsightsCommands.h"
 #include "UI/SSmithUEStatusIndicator.h"
 #include "Editor.h"
 #include "Modules/ModuleManager.h"
@@ -89,6 +90,7 @@ void FSmithUEModule::StartupModule()
 	FSmithUEDialogCommands::RegisterTools(FSmithUEToolRegistry::Get());
 	FSmithUEPCGCommands::RegisterTools(FSmithUEToolRegistry::Get());
 	FSmithUEAICommands::RegisterTools(FSmithUEToolRegistry::Get());
+	FSmithUEInsightsCommands::RegisterTools(FSmithUEToolRegistry::Get());
 
 	// Arm the modal-dialog watcher so blocking editor prompts can be observed and
 	// closed via worker-safe tools even while they jam the game thread.